fixed navbar styles and conflicting cdn
