+++ client/src/components/Account/style.css
@@ -1,5 +1,4 @@
 .profile-sm {
-    display:inline-block;
     width:300px;
     height:auto;
     max-width:80%;
