+++ client/src/components/FreelancerCard/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./style.css";
-import { Col, Image, Label } from "react-bootstrap";
+import { Image, Label } from "react-bootstrap";
 
 
 function FreelancerCard(props) {
